requirements.txt
python-dotenv==0.21.1
requests==2.31.0